requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.0
SQLAlchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.9
aiosqlite==0.20.0
requests==2.32.3
requests-cache==1.2.1
beautifulsoup4==4.12.3
tenacity==9.0.0
loguru==0.7.2
python-dotenv==1.0.1
celery==5.4.0
redis==5.0.8
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
scikit-learn==1.5.2
numpy==2.1.2
pandas==2.2.3
streamlit==1.39.0
httpx==0.27.2
geopy==2.4.1
ratelimit==2.2.1
playwright==1.47.0
